Rename: 로고 폴더 옮기고 해당 코드 수정
